UI refactoring
