+++ frontend/src/components/datatable/datatable.css
@@ -20,3 +20,7 @@ tbody tr:nth-of-type(odd) {
   flex-shrink: 0;
   margin-left: 20px;
 }
+
+.ida-card {
+  margin: 15px 0;
+}
