+++ frontend/src/components/home.css
@@ -1,0 +1,9 @@
+.nav-bar-container,
+.content {
+    max-height: 100vh;
+    overflow-y: auto;
+}
+
+.root {
+    flex-grow: 1;
+}
