+++ frontend/src/components/navbar/navbar.js
@@ -1,27 +1,11 @@
 import React from "react";
-import { makeStyles } from "@material-ui/core/styles";
 import TreeView from "@material-ui/lab/TreeView";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
 import ChevronRightIcon from "@material-ui/icons/ChevronRight";
 import TreeItem from "@material-ui/lab/TreeItem";
 import Box from "@material-ui/core/Box";
 
-const useStyles = makeStyles({
-  Box: {
-    height: 1000,
-    width: 450,
-    backgroundcolor: "lavender",
-  },
-
-  root: {
-    height: 216,
-    flexGrow: 1,
-    maxWidth: 400,
-  },
-});
-
 export default function RecursiveTreeView(props) {
-  const classes = useStyles();
   const detail = props.detail;
   const expanded = props.expandedNodeId;
   const chooseSelect = () => {
@@ -45,11 +29,9 @@ export default function RecursiveTreeView(props) {
   };
 
   return (
-
-    <div style={{ width: "100%", }}>
+    <div>
       <Box p={1} height="auto" >
         <TreeView
-          className={classes.root}
           defaultCollapseIcon={<ExpandMoreIcon />}
           expanded={expanded}
           defaultExpandIcon={<ChevronRightIcon />}
